backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
bandit==1.9.4
certifi==2026.1.4
charset-normalizer==3.4.4
cloudinary==1.44.1
colorama==0.4.6
decouple==0.0.7
dj-database-url==2.1.0
Django==6.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-environ==0.13.0
django-rest-framework==0.1.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
git-filter-repo==2.47.0
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.8.1
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pathspec==1.0.4
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pypdf==6.7.5
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
stevedore==5.7.0
tenseal==0.3.16
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
